new file miopen_db_utils.py added under miopen/utils  get_solver_id , get_id_solvers in new file
