+++ tuna/utils/db_utility.py
@@ -107,36 +107,6 @@ def create_indices(all_indices):
         continue
 
 
-def get_solver_ids():
-  """DB solver name to id map"""
-  # TODO: Get this info from the SQLAlchemy class  # pylint: disable=fixme
-  solver_id_map = {}
-  with DbSession() as session:
-    query = session.query(Solver.solver, Solver.id).filter(Solver.valid == 1)
-    res = session_retry(session, query.all, lambda x: x(), LOGGER)
-    for slv, sid in res:
-      solver_id_map[slv] = sid
-      solver_id_map[slv.replace(', ', '-')] = sid
-
-  return solver_id_map
-
-
-def get_id_solvers():
-  """DB solver id to name map"""
-  solver_id_map_c = {}
-  solver_id_map_h = {}
-  with DbSession() as session:
-    query = session.query(Solver.solver, Solver.id).filter(Solver.valid == 1)
-    res = session_retry(session, query.all, lambda x: x(), LOGGER)
-    for slv, sid in res:
-      solver_id_map_c[slv] = sid
-      solver_id_map_h[slv.replace(', ', '-')] = sid
-    id_solver_map_c = {val: key for key, val in solver_id_map_c.items()}
-    id_solver_map_h = {val: key for key, val in solver_id_map_h.items()}
-
-  return id_solver_map_c, id_solver_map_h
-
-
 def session_retry(session: DbSession,
                   callback: Callable,
                   actuator: Callable,
